refactor: reorganize services into domain folders
Move flat services/ into 7 domain folders: auth/, chat/, config/,
extensions/, genesis/, lens/, sdk/. Add barrel exports for each domain.
No logic changes — pure structural reorganization.

198 tests passing.
