+++ src/main.ts
@@ -2,23 +2,23 @@ import { app, BrowserWindow, ipcMain } from 'electron';
 import path from 'node:path';
 import * as fs from 'fs';
 import started from 'electron-squirrel-startup';
-import { ChatService } from './main/services/ChatService';
-import { ExtensionLoader } from './main/services/ExtensionLoader';
-import { ConfigService } from './main/services/ConfigService';
-import { AuthService } from './main/services/AuthService';
-import { MindScaffold } from './main/services/MindScaffold';
-import { IdentityLoader } from './main/services/IdentityLoader';
-import { seedLensDefaults, installLensSkill } from './main/services/MindBootstrap';
-import { loadCanvasExtension } from './main/services/adapters/canvas';
-import { loadCronExtension } from './main/services/adapters/cron';
-import { loadIdeaExtension } from './main/services/adapters/idea';
+import { ChatService } from './main/services/chat';
+import { ExtensionLoader } from './main/services/extensions';
+import { ConfigService } from './main/services/config';
+import { AuthService } from './main/services/auth';
+import { MindScaffold } from './main/services/genesis';
+import { IdentityLoader } from './main/services/chat';
+import { seedLensDefaults, installLensSkill } from './main/services/lens';
+import { loadCanvasExtension } from './main/services/extensions/adapters/canvas';
+import { loadCronExtension } from './main/services/extensions/adapters/cron';
+import { loadIdeaExtension } from './main/services/extensions/adapters/idea';
 import { setupChatIPC } from './main/ipc/chat';
 import { setupAgentIPC } from './main/ipc/agent';
 import { setupLensIPC } from './main/ipc/lens';
 import { setupGenesisIPC } from './main/ipc/genesis';
 import { setupAuthIPC } from './main/ipc/auth';
-import { stopSharedClient } from './main/services/SdkLoader';
-import { ViewDiscovery } from './main/services/ViewDiscovery';
+import { stopSharedClient } from './main/services/sdk';
+import { ViewDiscovery } from './main/services/lens';
 
 if (started) {
   app.quit();
