+++ src/main/services/auth/index.ts
@@ -1,0 +1,2 @@
+export { AuthService, getCredentialAccount, getLoginFromAccount, resolveStoredCredential } from './AuthService';
+export type { AuthProgress } from './AuthService';
