+++ src/main/services/chat/IdentityLoader.ts
