+++ src/main/services/chat/index.ts
@@ -1,0 +1,2 @@
+export { ChatService } from './ChatService';
+export { IdentityLoader } from './IdentityLoader';
