+++ src/main/services/extensions/ExtensionLoader.test.ts
