+++ src/main/services/extensions/adapters/canvas.ts
