+++ src/main/services/extensions/index.ts
@@ -1,0 +1,5 @@
+export { ExtensionLoader } from './ExtensionLoader';
+export type { ExtensionAdapter, ExtensionTool, LoadedExtension } from './ExtensionLoader';
+export { loadCanvasExtension } from './adapters/canvas';
+export { loadCronExtension } from './adapters/cron';
+export { loadIdeaExtension } from './adapters/idea';
