+++ src/main/services/genesis/MindScaffold.test.ts
