+++ src/main/services/genesis/MindScaffold.ts
@@ -4,7 +4,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
 import { execSync } from 'child_process';
-import { getSharedClient } from './SdkLoader';
+import { getSharedClient } from '../sdk/SdkLoader';
 import { buildGenesisPrompt } from './genesisPrompt';
 import { GitHubRegistryClient } from './GitHubRegistryClient';
