+++ src/main/services/genesis/genesisPrompt.ts
