+++ src/main/services/genesis/index.ts
@@ -1,0 +1,5 @@
+export { MindScaffold } from './MindScaffold';
+export type { GenesisConfig, GenesisProgress } from './MindScaffold';
+export { buildGenesisPrompt } from './genesisPrompt';
+export type { GenesisPromptInput } from './genesisPrompt';
+export { GitHubRegistryClient } from './GitHubRegistryClient';
