+++ src/main/services/lens/MindBootstrap.test.ts
