+++ src/main/services/lens/MindBootstrap.ts
