+++ src/main/services/lens/ViewDiscovery.test.ts
