+++ src/main/services/lens/index.ts
@@ -1,0 +1,2 @@
+export { ViewDiscovery } from './ViewDiscovery';
+export { seedLensDefaults, installLensSkill } from './MindBootstrap';
