+++ src/main/services/sdk/index.ts
@@ -1,0 +1,2 @@
+export { getSharedClient, stopSharedClient } from './SdkLoader';
+export { findSystemNode, requireSystemNode } from './nodeResolver';
