+++ src/main/services/sdk/nodeResolver.ts
